# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patient": {
      "get": {
        "operationId": "PatientController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role (OPERATIONS or SUPER_USER required for write operations)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all patients",
        "tags": [
          "Patients"
        ]
      },
      "post": {
        "operationId": "PatientController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role (OPERATIONS or SUPER_USER required for write operations)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient successfully created"
          }
        },
        "summary": "Register a new patient",
        "tags": [
          "Patients"
        ]
      }
    },
    "/patient/{id}": {
      "get": {
        "operationId": "PatientController_findOne",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role (OPERATIONS or SUPER_USER required for write operations)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a patient by ID",
        "tags": [
          "Patients"
        ]
      },
      "put": {
        "operationId": "PatientController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role (OPERATIONS or SUPER_USER required for write operations)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update patient information",
        "tags": [
          "Patients"
        ]
      },
      "delete": {
        "operationId": "PatientController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role (OPERATIONS or SUPER_USER required for write operations)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a patient",
        "tags": [
          "Patients"
        ]
      }
    },
    "/ward": {
      "get": {
        "operationId": "WardController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all wards",
        "tags": [
          "Wards"
        ]
      }
    },
    "/ward/{name}/beds": {
      "get": {
        "operationId": "WardController_findBeds",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get beds in a ward",
        "tags": [
          "Wards"
        ]
      }
    },
    "/ward/{name}/bed/{bedNumber}": {
      "put": {
        "operationId": "WardController_updateBed",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bedNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update bed status",
        "tags": [
          "Wards"
        ]
      }
    },
...
    "/request/pre": {
      "get": {
        "operationId": "RequestController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all PRE requests",
        "tags": [
          "Requests"
        ]
      },
      "post": {
        "operationId": "RequestController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePreRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new PRE request",
        "tags": [
          "Requests"
        ]
      }
    },
    "/request/pre/{id}": {
      "put": {
        "operationId": "RequestController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "User role",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a PRE request",
        "tags": [
          "Requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hospital Management System API",
    "description": "In-memory backend for Review-4",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role"
      }
    },
    "schemas": {
...
      "CreatePatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "FED-2026-9001"
          },
          "name": {
            "type": "string",
            "example": "Arun Mehta"
          },
          "age": {
            "type": "string",
            "example": "54"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "phone": {
            "type": "string",
            "example": "9876543210"
          },
          "address": {
            "type": "string",
            "example": "Banjara Hills"
          },
          "email": {
            "type": "string",
            "example": "arun.mehta@example.com"
          }
        },
        "required": [
          "name",
          "age",
          "gender",
          "phone",
          "address"
        ]
      },
...
      "CreatePreRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "PRE-TEST-2001"
          },
          "appointmentId": {
            "type": "string",
            "example": "PRE-APT-TEST-2001"
          },
          "patientId": {
            "type": "string",
            "example": "FED-2026-9001"
          },
          "name": {
            "type": "string",
            "example": "Arun Mehta"
          },
          "age": {
            "type": "string",
            "example": "54"
          },
          "gender": {
            "type": "string",
            "example": "Male"
          },
          "department": {
            "type": "string",
            "example": "Cardiology"
          },
          "status": {
            "type": "string",
            "example": "Pending",
            "default": "Pending"
          },
          "homStatus": {
            "type": "string",
            "example": "Awaiting HOM",
            "default": "Awaiting HOM"
          },
          "doctorId": {
            "type": "string",
            "example": "D101"
          },
          "date": {
            "type": "string",
            "example": "2026-05-10"
          },
          "timeSlot": {
            "type": "string",
            "example": "9:00 AM - 10:00 AM"
          },
          "problem": {
            "type": "string",
            "example": "Chest pain"
          }
        },
        "required": [
          "patientId",
          "name",
          "age",
          "gender",
          "department",
          "status",
          "homStatus",
          "doctorId",
          "date",
          "timeSlot",
          "problem"
        ]
      },
...
    }
  }
}